+++ setup.sh
@@ -7,7 +7,7 @@ echo "Creating dev environment in ./venv..."
 python3 -m venv venv
 . venv/bin/activate
 pip3 install -U pip setuptools
-python3 setup.py install
+pip3 install -r requirements.txt
 
 echo ""
 echo "  * Created virtualenv environment in ./venv."
