+++ shareinator/utils/scanner.py
